# backend_docs/swagger_api.json
{
    "swagger": "2.0",
    "info": {
      "title": "Reddit Clone API",
      "description": "API for Reddit Clone",
      "termsOfService": "https://www.example.com/terms/",
      "contact": {
        "email": "contact@example.com"
      },
      "license": {
        "name": "MIT License"
      },
      "version": "v1"
    },
    "host": "localhost:8000",
    "schemes": [
      "http"
    ],
    "basePath": "/api/v1",
    "consumes": [
      "application/json"
    ],
    "produces": [
      "application/json"
    ],
    "securityDefinitions": {
      "Basic": {
        "type": "basic"
      }
    },
    "security": [
      {
        "Basic": []
      }
    ],
    "paths": {
...
      "/posts/media/": {
        "get": {
          "operationId": "posts_media_list",
          "description": "API endpoint for post media.",
          "parameters": [
            {
              "name": "limit",
              "in": "query",
              "description": "Number of results to return per page.",
              "required": false,
              "type": "integer"
            },
            {
              "name": "offset",
              "in": "query",
              "description": "The initial index from which to return the results.",
              "required": false,
              "type": "integer"
            }
          ],
          "responses": {
            "200": {
              "description": "",
              "schema": {
                "required": [
                  "count",
                  "results"
                ],
                "type": "object",
                "properties": {
                  "count": {
                    "type": "integer"
                  },
                  "next": {
                    "type": "string",
                    "format": "uri",
                    "x-nullable": true
                  },
                  "previous": {
                    "type": "string",
                    "format": "uri",
                    "x-nullable": true
                  },
                  "results": {
                    "type": "array",
                    "items": {
                      "$ref": "#/definitions/PostMedia"
                    }
                  }
                }
              }
            }
          },
          "tags": [
            "posts"
          ]
        },
        "post": {
          "operationId": "posts_media_create",
          "description": "API endpoint for post media.",
          "parameters": [
            {
              "name": "data",
              "in": "body",
              "required": true,
              "schema": {
                "$ref": "#/definitions/PostMedia"
              }
            }
          ],
          "responses": {
            "201": {
              "description": "",
              "schema": {
                "$ref": "#/definitions/PostMedia"
              }
            }
          },
          "tags": [
            "posts"
          ]
        },
        "parameters": []
      },
...
    },
    "definitions": {
...
      "PostMedia": {
        "required": [
          "post",
          "media_type",
          "media_url"
        ],
        "type": "object",
        "properties": {
          "id": {
            "title": "Id",
            "type": "string",
            "format": "uuid",
            "readOnly": true
          },
          "post": {
            "title": "Post",
            "type": "string",
            "format": "uuid"
          },
          "media_type": {
            "title": "Media type",
            "type": "string",
            "maxLength": 50,
            "minLength": 1
          },
          "media_url": {
            "title": "Media url",
            "type": "string",
            "maxLength": 255,
            "minLength": 1
          },
          "thumbnail_url": {
            "title": "Thumbnail url",
            "type": "string",
            "maxLength": 255,
            "x-nullable": true
          },
          "order": {
            "title": "Order",
            "type": "integer",
            "maximum": 2147483647,
            "minimum": -2147483648
          },
          "created_at": {
            "title": "Created at",
            "type": "string",
            "format": "date-time",
            "readOnly": true
          }
        }
      },
...
    }
  }